// src/app/api/model/userDto.ts
/**
 * Rezflix Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserTypeDto } from './userTypeDto';


export interface UserDto { 
    about?: string;
    id: number;
    type: UserTypeDto;
    username: string;
}