// src/app/api/model/createUserDto.ts
/**
 * Rezflix Backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserTypeDto } from './userTypeDto';


export interface CreateUserDto { 
    about?: string;
    username: string;
    password: string;
    type: UserTypeDto;
}